+++ src/app/party/create/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import dynamic from 'next/dynamic';
 
-const PartyCreate = dynamic(() => import('@/components/pages/party/create/PartyCreate'), { ssr: false });
+const PartyCreate = dynamic(() => import('@/components/pages/party/edit/PartyEdit'), { ssr: false });
 
 export default function PartyCreatePage() {
   return (
