+++ src/app/party/modify/layout.tsx
@@ -1,6 +1,6 @@
 import type { PropsWithChildren } from 'react';
 import React from 'react';
 
-export default function PartyEditLayout({ children }: PropsWithChildren) {
+export default function PartyModifyLayout({ children }: PropsWithChildren) {
   return <main>{children}</main>;
 }
