+++ src/components/pages/party/recruit/PartyRecruit.tsx
@@ -102,22 +102,25 @@ function PartyRecruit() {
     openModal({
       children: (
         <ConfirmModal
-          modalTitle="경고!"
+          modalTitle="나가기"
           modalContents={
             <>
-              작성된 내용이 저장되지 않았습니다.
-              <br />이 페이지를 떠나시겠습니까?
+              파티 모집이 완료되지 않았어요.
+              <br />
+              입력하신 내용이 있으면 저장되지 않아요!
+              <br />
+              그래도 해당 페이지에서 정말 나가시겠어요?
             </>
           }
-          cancelBtnTxt="나가기"
-          submitBtnTxt="작성하기"
+          cancelBtnTxt="취소"
+          submitBtnTxt="나가기"
         />
       ),
       onCancel: () => {
-        router.push('/');
         closeModal();
       },
       onSubmit: () => {
+        router.push('/');
         closeModal();
       },
     });
@@ -128,7 +131,7 @@ function PartyRecruit() {
       children: (
         <ConfirmModal
           modalTitle="등록 완료"
-          modalContents={<div style={{}}>모집공고가 등록되었습니다.</div>}
+          modalContents={<div style={{}}>모집공고가 등록되었어요!</div>}
           submitBtnTxt="닫기"
         />
       ),
@@ -161,7 +164,7 @@ function PartyRecruit() {
 
   return (
     <SContainer>
-      <PageHeader title="파티원 모집하기" />
+      <PageHeader title="파티원 모집" />
       <PartyRecruitContainer>
         <SFlexColumnFull>
           <SMargin margin="3.125rem 0rem 0rem 0rem" />
